requirements.txt
transformers
gradio
scikit-learn
torch
matplotlib
python-docx
PyPDF2